+++ mbedtls-rs/src/lib.rs
@@ -31,6 +31,8 @@ pub use session::*;
 pub(crate) mod fmt; // MUST be the first so that the other modules can see it
 
 mod cert;
+#[cfg(feature = "ecp-restartable")]
+pub mod ecp;
 mod session;
 
 /// Re-export of the mbedtls-rs-sys crate so that users do not have to